separate C++14~ STL supplementations into stl_supplement.hpp
