+++ include/fkYAML/ordered_map.hpp
@@ -90,7 +90,7 @@ public:
      * @return mapped_type& Reference to a mapped_type object associated with the given key.
      */
     template <
-        typename KeyType, fkyaml::enable_if_t<is_usable_as_key_type<key_compare, key_type, KeyType>::value, int> = 0>
+        typename KeyType, detail::enable_if_t<is_usable_as_key_type<key_compare, key_type, KeyType>::value, int> = 0>
     mapped_type& operator[](KeyType&& key) noexcept
     {
         return emplace(std::forward<KeyType>(key), mapped_type()).first->second;
@@ -105,7 +105,7 @@ public:
      * @return std::pair<iterator, bool> A result of emplacement of the new key-value pair.
      */
     template <
-        typename KeyType, fkyaml::enable_if_t<is_usable_as_key_type<key_compare, key_type, KeyType>::value, int> = 0>
+        typename KeyType, detail::enable_if_t<is_usable_as_key_type<key_compare, key_type, KeyType>::value, int> = 0>
     std::pair<iterator, bool> emplace(KeyType&& key, const mapped_type& value) noexcept
     {
         for (auto itr = this->begin(); itr != this->end(); ++itr)
@@ -127,7 +127,7 @@ public:
      * @return mapped_type& The value associated to the given key.
      */
     template <
-        typename KeyType, fkyaml::enable_if_t<is_usable_as_key_type<key_compare, key_type, KeyType>::value, int> = 0>
+        typename KeyType, detail::enable_if_t<is_usable_as_key_type<key_compare, key_type, KeyType>::value, int> = 0>
     mapped_type& at(KeyType&& key)
     {
         for (auto itr = this->begin(); itr != this->end(); ++itr)
@@ -148,7 +148,7 @@ public:
      * @return const mapped_type& The value associated to the given key.
      */
     template <
-        typename KeyType, fkyaml::enable_if_t<is_usable_as_key_type<key_compare, key_type, KeyType>::value, int> = 0>
+        typename KeyType, detail::enable_if_t<is_usable_as_key_type<key_compare, key_type, KeyType>::value, int> = 0>
     const mapped_type& at(KeyType&& key) const
     {
         for (auto itr = this->begin(); itr != this->end(); ++itr)
@@ -169,7 +169,7 @@ public:
      * @return iterator The iterator for the found value, or the result of end().
      */
     template <
-        typename KeyType, fkyaml::enable_if_t<is_usable_as_key_type<key_compare, key_type, KeyType>::value, int> = 0>
+        typename KeyType, detail::enable_if_t<is_usable_as_key_type<key_compare, key_type, KeyType>::value, int> = 0>
     iterator find(KeyType&& key) noexcept
     {
         for (auto itr = this->begin(); itr != this->end(); ++itr)
@@ -190,7 +190,7 @@ public:
      * @return const_iterator The constant iterator for the found value, or the result of end().
      */
     template <
-        typename KeyType, fkyaml::enable_if_t<is_usable_as_key_type<key_compare, key_type, KeyType>::value, int> = 0>
+        typename KeyType, detail::enable_if_t<is_usable_as_key_type<key_compare, key_type, KeyType>::value, int> = 0>
     const_iterator find(KeyType&& key) const noexcept
     {
         for (auto itr = this->begin(); itr != this->end(); ++itr)
